requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-BasicAuth==0.2.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.2
python-zbase32==0.1.1
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.2.3
